pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = ["tests.py", "test_*.py"]

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"
custom_html = "c-\\w+,c-\\w+-\\w+,c-\\w+-\\w+-\\w+,c-\\w+-\\w+-\\w+-\\w+,c-\\w+-\\w+-\\w+-\\w+-\\w+,circle,rect,path,svg,use,line"
custom_blocks = "cotton,partialdef,slot"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = ["*/migrations/*.py", "staticfiles/*"]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  # of types for comparison.
  # Deactivated because it can make the code slow:
  # https://github.com/astral-sh/ruff/issues/7871
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
  "django-browser-reload==1.21.0",
  "django-debug-toolbar==6.1.0",
  "django-extensions==4.1",
  "djlint==1.36.4",
  "factory-boy==3.3.2",
  "psycopg[binary]==3.3.1",
  "pytest==9.0.2",
  "pytest-django==4.11.1",
  "pytest-sugar==1.1.1",
  "ruff==0.14.8",
  "watchfiles==1.1.1",
  "werkzeug[watchdog]==3.1.4",
]

[project]
name = "elastomeros-ecuatorianos"
version = "0.1.0"
description = "Elastomeros Ecuatorianos store"
readme = "README.md"
license = { text = "Not open source" }
authors = [{ name = "Dennis Paz", email = "dppazlopez@gmail.com" }]
requires-python = "==3.13.*"
dependencies = [
  "argon2-cffi==25.1.0",
  "celery==5.6.0",
  "crispy-bootstrap5==2025.6",
  "django==5.2.9",
  "django-allauth[mfa]==65.13.1",
  "django-anymail[amazon-ses]==13.1",
  "django-celery-beat==2.8.1",
  "django-braces==1.17.0",
  "django-cotton==2.5.0",
  "django-crispy-forms==2.5",
  "django-environ==0.12.0",
  "django-extra-views==0.16.0",
  "django-filter==25.2.0",
  "django-formtools==2.5.1",
  "django-htmx==1.27.0",
  "django-model-utils==5.0.0",
  "django-nested-admin==4.1.6",
  "django-phonenumber-field[phonenumberslite]==8.4.0",
  "django-redis==6.0.0",
  "django-storages[s3]==1.14.6",
  "django-template-partials==25.3",
  "django-widget-tweaks==1.5.0",
  "gunicorn==23.0.0",
  "hiredis==3.3.0",
  "pillow==12.0.0",
  "psycopg[c]==3.3.1",
  "python-slugify==8.0.4",
  "redis==7.1.0",
  "whitenoise==6.11.0",
]
